fix(identity): move password helpers out of service.go
Keep service.go under the 1200-line structure guard while preserving
the strengthened password policy and auto-generate behavior.
